LoginComponent tests mock hooks
