+++ apps/tup-ui/src/auth/LoginComponent.spec.tsx
@@ -2,67 +2,31 @@ import { render, screen } from '@testing-library/react';
 import { AxiosStatic } from 'axios';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import LoginComponent, {
-  ProfileComponent,
-} from './LoginComponent';
+import LoginComponent from './LoginComponent';
+import useAuth from '../hooks/useAuth';
+import useProfile from '../hooks/useProfile';
 
-import { AxiosClientContext } from '../hooks/useAxios';
+jest.mock('../hooks/useAuth');
+jest.mock('./ProfileComponent');
 
-
-
-const mockedAxios = jest.createMockFromModule<AxiosStatic>('axios');
-const testQueryClient = new QueryClient({
-  defaultOptions: { queries: { retry: false, cacheTime: Infinity } },
-});
-const WrappedProfileComponent = () => (
-  <AxiosClientContext.Provider value={mockedAxios}>
-    <QueryClientProvider client={testQueryClient}>
-      <ProfileComponent />
-    </QueryClientProvider>
-  </AxiosClientContext.Provider>
-);
-
-const WrappedLogin = () => (
-  <AxiosClientContext.Provider value={mockedAxios}>
-    <QueryClientProvider client={testQueryClient}>
-      <LoginComponent />
-    </QueryClientProvider>
-  </AxiosClientContext.Provider>
-);
 
 describe('LoginComponent', () => {
-  afterEach(() => testQueryClient.clear());
-  it('should render profile if cookie is defined', async () => {
-    document.cookie = 'x-tup-token=abc123';
-    jest
-      .spyOn(mockedAxios, 'get')
-      .mockResolvedValue({ data: { firstName: 'testuser' } });
-    render(<WrappedLogin />);
-
-    const userQuery = await screen.findByText(/testuser/);
-    expect(userQuery).toBeTruthy();
+  it('should render login component if not logged in', () => {
+    (useAuth as jest.Mock).mockReturnValue({
+      loggedIn: false,
+      isLoading: false,
+      error: null
+    })
+    const { getAllByText } = render(<LoginComponent />);
+    expect(getAllByText(/Login/)).toBeTruthy();
   });
-
-  it('should render login page if no cookie is set', async () => {
-    document.cookie = 'x-tup-token=';
-
-    render(<WrappedLogin />);
-
-    const userQuery = await screen.findByText(/Login/);
-    expect(userQuery).toBeTruthy();
-  });
-});
-
-describe('ProfileComponent', () => {
-  afterEach(() => testQueryClient.clear());
-  it('should render successfully', async () => {
-    jest
-      .spyOn(mockedAxios, 'get')
-      .mockResolvedValue({ data: { firstName: 'testuser' } });
-
-    render(<WrappedProfileComponent />);
-    expect(mockedAxios.get).toHaveBeenCalled();
-    const userQuery = await screen.findByText(/testuser/);
-    expect(userQuery).toBeTruthy();
+  it('should render a logout component if logged in', () => {
+    (useAuth as jest.Mock).mockReturnValue({
+      loggedIn: true,
+      isLoading: false,
+      error: null
+    })
+    const { getAllByText } = render(<LoginComponent />);
+    expect(getAllByText(/Log Out/)).toBeTruthy();
   });
 });
