+++ apps/tup-ui/src/hooks/useJwt.ts
@@ -4,7 +4,7 @@ import Cookies from 'js-cookie';
 //Retrieve a JWT stored in the x-tup-token cookie.
 const getJwt = () => {
   const cookie = Cookies.get('x-tup-token');
-  return cookie; 
+  return cookie;
 };
 
 const useJwt = () => {
